use TLSv1 by default and fallback to SSLv3 when **needed**
The library used to fail to connect to a server which does not
permit SSLv3.

With this patch, by default, TLSv1 is tried. If it's not available
the library downgrades to SSLv3.

Even though this is unwanted behaviour, this patch should allows
the client to be used with a server that does not allow SSLv3
while not breaking backwards compatibility.